core: change handles to a more suitable type (#477)
Description: Better fit for storing opaque memory addresses and/or sentinel values.

+++ mobile/library/swift/test/MockEnvoyHTTPStream.swift
@@ -6,7 +6,7 @@ final class MockEnvoyHTTPStream {
   static var onData: ((Data, Bool) -> Void)?
   static var onTrailers: (([String: [String]]) -> Void)?
 
-  init(handle: UInt64, callbacks: EnvoyHTTPCallbacks) {}
+  init(handle: Int, callbacks: EnvoyHTTPCallbacks) {}
 }
 
 extension MockEnvoyHTTPStream: EnvoyHTTPStream {
